Introducing ResourceLinks, and other changes
ResourceLinks are a way to provide a generic REST-ful interface to
an arbitrary resource. They provide a way for automation primitives
(like AutomationPairings and AutomationRules) to perform REST-ful
operations without knowing any details about the underlying transport.

This change also changes how the LocalTechnology choose a native
URI path for host functional endpoints: it now does so using an
integer instead of a UID. This allows the UID to be changed without
changing the path of the FE.