ResourceLink Refactoring
Since ResourceLinks are always managed, this change adds a new method
to the ResourceLink class to get the URI. As a result, the method
`Technology#getNativeUriForResourceLink()` has been removed since it
is now redundant.

Spun off the method `getResourceLinkForNativeUri()` to its own
interface definition. Decoupling this functionality from the
Technology class will be helpful for future work.

These are resource links that can't be immediately resolved when
they are requested, but may resolve later once everything has loaded.
Will also eventually be used to provide some indirection for UID
lookups.