LocalTechnology: Cache ResourceLinks for a given URI
This change sets up a cache of ResourceLinks that have been looked up
by the LocalTechnology. This cache is weak, so GC can collect
ResourceLinks which are not referenced. Until the GC collects the
ResourceLink it can still be looked up and returned by
`getResourceLinkForNativeUri()`. 

`PropertyResourceLink` and `SectionResourceLink` have been updated to
keep a reference to the URI itself, which simplifies cache bookkeeping
since the key will be collected along with the ResourceLink.

This is similar to the ResourceLink cache implemented in SmcpTechnology.